@@ -12,12 +12,13 @@
     "build": "node build/build.js",
     "unit": "karma start test/unit/karma.conf.js --single-run",
     "test": "npm run unit",
-    "lint": "eslint --ext .js src test/unit/specs"
+    "lint": "eslint --ext .js src"
   },
   "keywords": [],
   "dependencies": {
     "antd-mobile": "^0.9.14",
     "classnames": "^2.2.5",
+    "core-decorators": "^0.15.0",
     "immutable": "^3.8.1",
     "rc-tabs": "^7.1.0",
     "react": "^15.1.0",
@@ -31,12 +32,13 @@
     "redux-saga": "^0.13.0",
     "reduxsauce": "^0.4.1",
     "reselect": "^2.5.4",
-    "sanitize.css": "^4.1.0"
+    "sanitize.css": "^4.1.0",
+    "whatwg-fetch": "^2.0.1"
   },
   "devDependencies": {
     "autoprefixer": "^6.4.0",
     "babel-core": "^6.0.0",
-    "babel-eslint": "^7.0.0",
+    "babel-eslint": "^7.1.1",
     "babel-loader": "^6.0.0",
     "babel-plugin-import": "^1.1.0",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
@@ -49,19 +51,21 @@
     "chalk": "^1.1.3",
     "cheerio": "^0.22.0",
     "connect-history-api-fallback": "^1.1.0",
-    "core-decorators": "^0.15.0",
     "css-loader": "^0.25.0",
-    "eslint": "^3.7.1",
+    "eslint": "^3.13.1",
+    "eslint-config-airbnb": "^14.0.0",
     "eslint-config-standard": "^6.1.0",
     "eslint-friendly-formatter": "^2.0.5",
-    "eslint-loader": "^1.5.0",
+    "eslint-loader": "^1.6.1",
     "eslint-plugin-html": "^1.3.0",
+    "eslint-plugin-import": "^2.2.0",
+    "eslint-plugin-jsx-a11y": "^3.0.2",
     "eslint-plugin-promise": "^3.4.0",
+    "eslint-plugin-react": "^6.9.0",
     "eslint-plugin-standard": "^2.0.1",
     "eventsource-polyfill": "^0.9.6",
     "express": "^4.13.3",
     "extract-text-webpack-plugin": "^1.0.1",
-    "fecs-loader": "^0.3.1",
     "file-loader": "^0.9.0",
     "function-bind": "^1.0.2",
     "highlight.js": "^9.9.0",
