@@ -3,11 +3,9 @@
  * @author yankun01
  */
 
-import Types from '../constants/Global'
+import Types from '../constants/Global';
 
-export const setCurrentUser = (user) => {
-    return {
-        type: Types.SET_CURRENT_USER,
-        payload: user
-    }
-}
+export const setCurrentUser = user => ({
+  type: Types.SET_CURRENT_USER,
+  payload: user,
+});
