@@ -1,23 +1,24 @@
-import React, { Component, PropTypes } from 'react'
+import React, { Component, PropTypes } from 'react';
 
 export default class Todo extends Component {
 
-    static propTypes = {
-        onClick: PropTypes.func.isRequired,
-        text: PropTypes.string.isRequired,
-        completed: PropTypes.bool.isRequired
-    }
+  static propTypes = {
+    onClick: PropTypes.func.isRequired,
+    text: PropTypes.string.isRequired,
+    completed: PropTypes.bool.isRequired,
+  }
 
-    render() {
-        return (
-            <li
-                onClick={this.props.onClick}
-                style={{
-                    textDecoration: this.props.completed ? 'line-through' : 'none',
-                    cursor: this.props.completed ? 'default' : 'pointer'
-                }}>
-                {this.props.text}
-            </li>
-        )
-    }
+  render() {
+    return (
+      <li
+        onClick={this.props.onClick}
+        style={{
+          textDecoration: this.props.completed ? 'line-through' : 'none',
+          cursor: this.props.completed ? 'default' : 'pointer',
+        }}
+      >
+        {this.props.text}
+      </li>
+    );
+  }
 }
