@@ -1,26 +1,27 @@
-import React, { Component, PropTypes } from 'react'
-import Todo from './Todo'
+import React, { Component, PropTypes } from 'react';
+import Todo from './Todo';
 
 export default class TodoList extends Component {
 
-    static propTypes = {
-        onTodoClick: PropTypes.func.isRequired,
-        todos: PropTypes.arrayOf(PropTypes.shape({
-            text: PropTypes.string.isRequired,
-            completed: PropTypes.bool.isRequired
-        }).isRequired).isRequired
-    }
+  static propTypes = {
+    onTodoClick: PropTypes.func.isRequired,
+    todos: PropTypes.arrayOf(PropTypes.shape({
+      text: PropTypes.string.isRequired,
+      completed: PropTypes.bool.isRequired,
+    }).isRequired).isRequired,
+  }
 
-    render() {
-        return (
-            <ul>
-                {this.props.todos.map((todo, index) =>
-                    <Todo {...todo}
-                        key={index}
-                        onClick={() => this.props.onTodoClick(index)}
-                    />
-                )}
-            </ul>
-        )
-    }
+  render() {
+    return (
+      <ul>
+        {this.props.todos.map((todo, index) =>
+          <Todo
+            {...todo}
+            key={todo.id}
+            onClick={() => this.props.onTodoClick(index)}
+          />,
+        )}
+      </ul>
+    );
+  }
 }
