@@ -3,16 +3,16 @@
  * @author yankun01
  */
 
-import { createTypes } from 'reduxsauce'
+import { createTypes } from 'reduxsauce';
 
 export default createTypes(`
-    GET_LIST_DATA
-    GET_LIST_DATA_REQUEST
-    GET_LIST_DATA_REQUEST_SUCCESS
-    GET_LIST_DATA_REQUEST_FAILURE
+  GET_LIST_DATA
+  GET_LIST_DATA_REQUEST
+  GET_LIST_DATA_REQUEST_SUCCESS
+  GET_LIST_DATA_REQUEST_FAILURE
 
-    GET_LIST_ITEM_DATA
-    GET_LIST_ITEM_DATA_REQUEST
-    GET_LIST_ITEM_DATA_REQUEST_SUCCESS
-    GET_LIST_ITEM_DATA_REQUEST_FAILURE
-`)
+  GET_LIST_ITEM_DATA
+  GET_LIST_ITEM_DATA_REQUEST
+  GET_LIST_ITEM_DATA_REQUEST_SUCCESS
+  GET_LIST_ITEM_DATA_REQUEST_FAILURE
+`);
