@@ -7,11 +7,11 @@
 import React from 'react';
 
 export default function NotFound() {
-    return (
-        <article>
-            <h1>
-                404 Not Found
-            </h1>
-        </article>
-    );
+  return (
+    <article>
+      <h1>
+        404 Not Found
+      </h1>
+    </article>
+  );
 }
