@@ -3,9 +3,8 @@
  * @author maoquan
  */
 
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
-import BaseList from '../Base/BaseList'
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import { ListView } from 'antd-mobile';
 
 const data = [
@@ -31,120 +30,109 @@ const NUM_ROWS = 20;
 let pageIndex = 0;
 
 const genData = (pIndex = 0) => {
-    const dataBlob = {};
-    for (let i = 0; i < NUM_ROWS; i++) {
-        const ii = (pIndex * NUM_ROWS) + i;
-        dataBlob[`${ii}`] = `row - ${ii}`;
-    }
-    return dataBlob;
+  const dataBlob = {};
+  for (let i = 0; i < NUM_ROWS; i++) {
+    const ii = (pIndex * NUM_ROWS) + i;
+    dataBlob[`${ii}`] = `row - ${ii}`;
+  }
+  return dataBlob;
 };
 
-class ProductList extends Component {
+export default class ProductList extends Component {
 
-    static propTypes = {
+  static propTypes = {
 
-    }
+  }
 
-    constructor(props) {
-        super(props);
+  constructor(props) {
+    super(props);
 
-        const dataSource = new ListView.DataSource({
-            rowHasChanged: (row1, row2) => row1 !== row2,
-        });
-        this.rData = genData();
-        this.state = {
-            dataSource: dataSource.cloneWithRows(this.rData),
-            isLoading: false,
-        };
-    }
-
-    componentDidMount() {
-        // you can scroll to the specified position
-        // this.refs.lv.refs.listview.scrollTo(0, 200);
-    }
-
-    onEndReached(event) {
-        // load new data
-        console.log('reach end', event);
-        this.setState({ isLoading: true });
-        setTimeout(() => {
-            this.rData = { ...this.rData, ...genData(++pageIndex) };
-            this.setState({
-                dataSource: this.state.dataSource.cloneWithRows(this.rData),
-                isLoading: false,
-            });
-        }, 1000);
-    }
-
-    render() {
-        const separator = (sectionID, rowID) => (
-            <div
-                key={`${sectionID}-${rowID}`}
-                style={{
-                    backgroundColor: '#F5F5F9',
-                    height: 8,
-                    borderTop: '1px solid #ECECED',
-                    borderBottom: '1px solid #ECECED',
-                }}
-            />
-        );
-        const row = (rowData, sectionID, rowID) => {
-            if (index < 0) {
-                index = data.length - 1;
-            }
-            const obj = data[index--];
-            return (
-                <div key={rowID}
-                    style={{
-                        padding: '0.08rem 0.16rem',
-                        backgroundColor: 'white',
-                    }}
-                >
-                    <h3 style={{ padding: 2, marginBottom: '0.08rem', borderBottom: '1px solid #F6F6F6' }}>
-                        {obj.title}
-                    </h3>
-                    <div style={{ display: '-webkit-box', display: 'flex' }}>
-                        <img style={{ height: '1.28rem', marginRight: '0.08rem' }} src={obj.img} />
-                        <div style={{ display: 'inline-block' }}>
-                            <p>{obj.des}</p>
-                            <p><span style={{ fontSize: '1.6em', color: '#FF6E27' }}>{rowID}</span>元/任务</p>
-                        </div>
-                    </div>
-                </div>
-            );
-        };
-        return (
-            <ListView
-                dataSource={this.state.dataSource}
-                renderHeader={() => <span>header</span>}
-                renderFooter={
-                    () => <div style={{ padding: 30, textAlign: 'center' }}>
-                        {this.state.isLoading ? '加载中...' : '加载完毕'}
-                    </div>
-                }
-                renderRow={row}
-                renderSeparator={separator}
-                className="am-list"
-                pageSize={4}
-                scrollRenderAheadDistance={500}
-                scrollEventThrottle={20}
-                onScroll={() => { console.log('scroll'); }}
-                useBodyScroll
-                onEndReached={this.onEndReached}
-                onEndReachedThreshold={10}
-            />
-        );
-    }
-}
-
-const mapStateToProps = (state, ownProps) => {
-    return {
+    const dataSource = new ListView.DataSource({
+      rowHasChanged: (row1, row2) => row1 !== row2,
+    });
+    this.rData = genData();
+    this.state = {
+      dataSource: dataSource.cloneWithRows(this.rData),
+      isLoading: false,
     };
-}
+  }
 
-const mapDispatchToProps = (dispatch) => {
-    return {
+  componentDidMount() {
+    // you can scroll to the specified position
+    // this.refs.lv.refs.listview.scrollTo(0, 200);
+  }
+
+  onEndReached(event) {
+    // load new data
+    console.log('reach end', event);
+    this.setState({ isLoading: true });
+    setTimeout(() => {
+      this.rData = { ...this.rData, ...genData(++pageIndex) };
+      this.setState({
+        dataSource: this.state.dataSource.cloneWithRows(this.rData),
+        isLoading: false,
+      });
+    }, 1000);
+  }
+
+  render() {
+    const separator = (sectionID, rowID) => (
+      <div
+        key={`${sectionID}-${rowID}`}
+        style={{
+          backgroundColor: '#F5F5F9',
+          height: 8,
+          borderTop: '1px solid #ECECED',
+          borderBottom: '1px solid #ECECED',
+        }}
+      />
+    );
+    const row = (rowData, sectionID, rowID) => {
+      if (index < 0) {
+        index = data.length - 1;
+      }
+      const obj = data[index--];
+      return (
+        <div
+          key={rowID}
+          style={{
+            padding: '0.08rem 0.16rem',
+            backgroundColor: 'white',
+          }}
+        >
+          <h3 style={{ padding: 2, marginBottom: '0.08rem', borderBottom: '1px solid #F6F6F6' }}>
+            {obj.title}
+          </h3>
+          <div style={{ display: 'flex' }}>
+            <img style={{ height: '1.28rem', marginRight: '0.08rem' }} src={obj.img} alt={obj.title} />
+            <div style={{ display: 'inline-block' }}>
+              <p>{obj.des}</p>
+              <p><span style={{ fontSize: '1.6em', color: '#FF6E27' }}>{rowID}</span>元/任务</p>
+            </div>
+          </div>
+        </div>
+      );
     };
+    return (
+      <ListView
+        dataSource={this.state.dataSource}
+        renderHeader={() => <span>header</span>}
+        renderFooter={
+          () => <div style={{ padding: 30, textAlign: 'center' }}>
+            {this.state.isLoading ? '加载中...' : '加载完毕'}
+          </div>
+        }
+        renderRow={row}
+        renderSeparator={separator}
+        className="am-list"
+        pageSize={4}
+        scrollRenderAheadDistance={500}
+        scrollEventThrottle={20}
+        onScroll={() => { console.log('scroll'); }}
+        useBodyScroll
+        onEndReached={this.onEndReached}
+        onEndReachedThreshold={10}
+      />
+    );
+  }
 }
-
-export default connect(mapStateToProps, mapDispatchToProps)(ProductList)
