@@ -3,40 +3,27 @@
  * @author maoquan
  */
 
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
-import { ListView } from 'antd-mobile';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
 
-class CustomList extends Component {
+export default class CustomList extends Component {
 
-    static propTypes = {
+  static propTypes = {
 
-    }
+  }
 
-    constructor(props) {
-        super(props);
-    }
+  constructor(props) {
+    super(props);
+  }
 
-    componentDidMount() {
-        // you can scroll to the specified position
-        // this.refs.lv.refs.listview.scrollTo(0, 200);
-    }
+  componentDidMount() {
+    // you can scroll to the specified position
+    // this.refs.lv.refs.listview.scrollTo(0, 200);
+  }
 
-    render() {
-        return (
-            <p>报表</p>
-        );
-    }
+  render() {
+    return (
+      <p>报表</p>
+    );
+  }
 }
-
-const mapStateToProps = (state, ownProps) => {
-    return {
-    };
-}
-
-const mapDispatchToProps = (dispatch) => {
-    return {
-    };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(CustomList)
