@@ -22,34 +22,34 @@ import listReducer from './listReducer';
 
 // Initial routing state
 const routeInitialState = fromJS({
-    locationBeforeTransitions: null,
+  locationBeforeTransitions: null,
 });
 
 /**
- * Merge route into the global application state
- *
- * @return {Object}
- */
+* Merge route into the global application state
+*
+* @return {Object}
+*/
 function routeReducer(state = routeInitialState, action) {
-    switch (action.type) {
-        /* istanbul ignore next */
-        case LOCATION_CHANGE:
-            return state.merge({
-                locationBeforeTransitions: action.payload,
-            });
-        default:
-            return state;
-    }
+  switch (action.type) {
+    /* istanbul ignore next */
+    case LOCATION_CHANGE:
+      return state.merge({
+        locationBeforeTransitions: action.payload,
+      });
+    default:
+      return state;
+  }
 }
 
 /**
 * Creates the main reducer with the asynchronously loaded ones
 */
 export default function createReducer(asyncReducers) {
-    return combineReducers({
-        route: routeReducer,
-        todo: todoReducer,
-        list: listReducer,
-        ...asyncReducers,
-    });
+  return combineReducers({
+    route: routeReducer,
+    todo: todoReducer,
+    list: listReducer,
+    ...asyncReducers,
+  });
 }
