@@ -1,7 +1,7 @@
-import { fork } from 'redux-saga/effects'
-import listSaga from './listSaga'
-import request from '../utils/request'
+import { fork } from 'redux-saga/effects';
+import listSaga from './listSaga';
+import request from '../utils/request';
 
-export default function * root () {
-    yield fork(listSaga(request).watcher)
+export default function* root() {
+  yield fork(listSaga(request).watcher);
 }
