@@ -8,55 +8,55 @@ import { fromJS } from 'immutable';
 import { routerMiddleware } from 'react-router-redux';
 import createSagaMiddleware from 'redux-saga';
 import createReducer from './reducers';
-import sagas from './sagas'
+import sagas from './sagas';
 
 const sagaMiddleware = createSagaMiddleware();
 
 export default function configureStore(initialState = {}, history) {
-    // Create the store with two middlewares
-    // 1. sagaMiddleware: Makes redux-sagas work
-    // 2. routerMiddleware: Syncs the location/URL path to the state
-    const middlewares = [
-        sagaMiddleware,
-        routerMiddleware(history),
-    ];
+  // Create the store with two middlewares
+  // 1. sagaMiddleware: Makes redux-sagas work
+  // 2. routerMiddleware: Syncs the location/URL path to the state
+  const middlewares = [
+    sagaMiddleware,
+    routerMiddleware(history),
+  ];
 
-    const enhancers = [
-        applyMiddleware(...middlewares),
-    ];
+  const enhancers = [
+    applyMiddleware(...middlewares),
+  ];
 
-    // If Redux DevTools Extension is installed use it, otherwise use Redux compose
-    /* eslint-disable no-underscore-dangle */
-    const composeEnhancers
-        = process.env.NODE_ENV !== 'production'
-        && typeof window === 'object'
-        && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
-            ?  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ : compose;
+  // If Redux DevTools Extension is installed use it, otherwise use Redux compose
+  /* eslint-disable no-underscore-dangle */
+  const composeEnhancers
+    = process.env.NODE_ENV !== 'production'
+    && typeof window === 'object'
+    && window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__
+      ? window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ : compose;
 
-    const store = createStore(
-        createReducer(),
-        fromJS(initialState),
-        composeEnhancers(...enhancers)
-    );
+  const store = createStore(
+    createReducer(),
+    fromJS(initialState),
+    composeEnhancers(...enhancers),
+  );
 
-    // Extensions
-    // store.runSaga = sagaMiddleware.run;
-    store.asyncReducers = {}; // Async reducer registry
+  // Extensions
+  // store.runSaga = sagaMiddleware.run;
+  store.asyncReducers = {}; // Async reducer registry
 
-    // Make reducers hot reloadable, see http://mxs.is/googmo
-    /* istanbul ignore next */
-    if (module.hot) {
-        module.hot.accept('./reducers', () => {
-            System.import('./reducers').then((reducerModule) => {
-                const createReducers = reducerModule.default;
-                const nextReducers = createReducers(store.asyncReducers);
+  // Make reducers hot reloadable, see http://mxs.is/googmo
+  /* istanbul ignore next */
+  if (module.hot) {
+    module.hot.accept('./reducers', () => {
+      System.import('./reducers').then((reducerModule) => {
+        const createReducers = reducerModule.default;
+        const nextReducers = createReducers(store.asyncReducers);
 
-                store.replaceReducer(nextReducers);
-            });
-        });
-    }
+        store.replaceReducer(nextReducers);
+      });
+    });
+  }
 
-    sagaMiddleware.run(sagas)
+  sagaMiddleware.run(sagas);
 
-    return store;
+  return store;
 }
